move the canDowngrade method to a new helper (Account Helper)
