+++ resources/views/settings/subscriptions/downgrade-checklist.blade.php
@@ -63,7 +63,7 @@
           <form method="POST" action="{{ route('settings.subscriptions.downgrade') }}">
             @csrf
 
-            @if (auth()->user()->account->canDowngrade())
+            @if ($canDowngrade)
             <p class="mb4"><button href="" class="btn btn-primary">{{ trans('settings.subscriptions_downgrade_cta') }}</button></p>
             @else
             <p class="mb4"><button class="btn btn-primary" disabled="disabled">{{ trans('settings.subscriptions_downgrade_cta') }}</button></p>
