+++ tests/Unit/Models/AccountTest.php
@@ -254,73 +254,6 @@ class AccountTest extends FeatureTestCase
     }
 
     /** @test */
-    public function user_can_downgrade_with_only_one_user_and_no_pending_invitations_and_under_contact_limit()
-    {
-        config(['monica.number_of_allowed_contacts_free_account' => 1]);
-        $contact = factory(Contact::class)->create();
-        $account = $contact->account;
-
-        $user = factory(User::class)->create([
-            'account_id' => $account->id,
-        ]);
-
-        $this->assertEquals(
-            true,
-            $account->canDowngrade()
-        );
-    }
-
-    /** @test */
-    public function user_cant_downgrade_with_two_users()
-    {
-        $contact = factory(Contact::class)->create();
-        $account = $contact->account;
-
-        $user = factory(User::class)->create([
-            'account_id' => $account->id,
-        ]);
-
-        $user = factory(User::class)->create([
-            'account_id' => $account->id,
-        ]);
-
-        $this->assertEquals(
-            false,
-            $account->canDowngrade()
-        );
-    }
-
-    /** @test */
-    public function user_cant_downgrade_with_pending_invitations()
-    {
-        $account = factory(Account::class)->create();
-
-        $invitation = factory(Invitation::class)->create([
-            'account_id' => $account->id,
-        ]);
-
-        $this->assertEquals(
-            false,
-            $account->canDowngrade()
-        );
-    }
-
-    /** @test */
-    public function user_cant_downgrade_with_too_many_contacts()
-    {
-        config(['monica.number_of_allowed_contacts_free_account' => 1]);
-        $account = factory(Account::class)->create();
-
-        $contact = factory(Contact::class, 2)->create([
-            'account_id' => $account->id,
-        ]);
-
-        $this->assertFalse(
-            $account->canDowngrade()
-        );
-    }
-
-    /** @test */
     public function user_is_subscribed_if_user_can_access_to_paid_version_for_free()
     {
         $account = factory(Account::class)->make([
